refactor #step2 - Set<Integer> 대신 Lottery 클래스를 사용합니다
